Extract SaveFilesDialog and PhasingManual from main_window
