+++ main_window.py
@@ -13,7 +13,6 @@ import pyqtgraph as pg
 from pyqtgraph import mkPen, ColorMap, mkColor, InfiniteLine
 import pyqtgraph.exporters
 from ui_Form import Ui_NMR
-from ui_PhasingManual import Ui_Phasing as Ui_PhasingManual
 import Calculator as Cal # Mathematical procedures
 from controllers import (
     SETabController, DQTabController, DQTempTabController,
@@ -23,6 +22,8 @@ from dialogs.open_files_dialog import OpenFilesDialog
 import dialogs.open_files_dialog as open_files_dialog_module
 from dialogs.notification_dialog import NotificationDialog
 from dialogs.group_window import GroupWindow
+from dialogs.save_files_dialog import SaveFilesDialog
+from dialogs.phasing_manual import PhasingManual
 from widgets.table_copy_enabler import TableCopyEnabler
 from app_state import AppState
 
@@ -1166,260 +1167,3 @@ class MainWindow(QMainWindow):
         QCoreApplication.processEvents()
         return Fur
 
-class SaveFilesDialog(QFileDialog):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setAcceptMode(QFileDialog.AcceptSave)  # Set the dialog to save mode
-
-    def save_data_as_csv(self, directory, table, files, default_filename):
-        # I have no fucjing idea, why the hell this function takes these arguments, but it doesn't work otherwise.
-        try:
-            key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, r"Software\MyApp", 0, winreg.KEY_READ)
-            directory, _ = winreg.QueryValueEx(key, "SelectedFolder")
-            winreg.CloseKey(key)
-        except FileNotFoundError:
-            return os.path.dirname(sys.argv[0])
-        except Exception as e:
-            print(f"Couldn't read the initial directory: {e}")
-            return os.path.dirname(sys.argv[0])
-
-
-        options = QFileDialog.Options()
-        file_path, _ = QFileDialog.getSaveFileName(self, "Save File As", directory + '/' + default_filename, "CSV files (*.csv)", options=options)
-
-        if file_path:
-            try:
-                with open(file_path, 'w') as f:
-                    for row in range(table.rowCount()):
-                        row_values = []
-                        for col in range(table.columnCount()):
-                            item = table.item(row, col)
-                            if item is not None:
-                                row_values.append(item.text())
-                            else:
-                                row_values.append("")  # Handle empty cells
-                        f.write(','.join(row_values) + '\n')
-
-                files_list_path = os.path.splitext(file_path)[0] + '_files.json'
-
-                with open(files_list_path, 'w') as file_list:
-                    json.dump(files, file_list)
-
-            except Exception as e:
-                print(f"Failed to save file as CSV: {e}")
-
-    def save_file_in_sef(self, wtf, dictionary, tau, n, begin, save):
-        if save == False:
-            return
-
-        try:
-            key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, r"Software\MyApp", 0, winreg.KEY_READ)
-            directory, _ = winreg.QueryValueEx(key, "SelectedFolder")
-            winreg.CloseKey(key)
-        except FileNotFoundError:
-            return os.path.dirname(sys.argv[0])
-        except Exception as e:
-            print(f"Couldn't read the initial directory: {e}")
-            return os.path.dirname(sys.argv[0])
-
-        try:
-            options = QFileDialog.Options()
-            default_add = '_recalculated_tau_' + str(n)
-            file_path, _ = QFileDialog.getSaveFileName(self, "Save File As", directory + '/'+ begin + default_add, "SEF files (*.sef)", options=options)
-            with open(file_path, 'w') as f:
-                f.write('STELAR Export File\n')  
-                f.write('\n')  
-                f.write('_BRLX______\t_T1________\t_R1________\t____%err___\t+-err(R1)__\tZone\tFile\n')  
-                f.write('\n')
-                for key in dictionary:
-                    Frequency = dictionary[key]['X Axis']
-                    T1 = dictionary[key][tau]
-                    if T1 == 0:
-                        Omega_1 = 0
-                    else:
-                        Omega_1 = 1000/T1
-                    f.write(f'{Frequency}\t{T1}\t{Omega_1}\n')
-        except:
-            QMessageBox.warning(self, "Save failed", f"Sorry, couldn't save the separate file in .sef for magnetization.", QMessageBox.Ok)
-            return
-        #TODO: try to find reasons for that maybe? 
-        # When the load happends, there is no key T1 1 created in dictionary
-        # So ppb should read value from the table
-
-class PhasingManual(QDialog):
-    # TODO somehow create an araay of ORIGINAL Re_spectra and be able to restore it...
-    closed = Signal()
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.ui = Ui_PhasingManual()
-        self.ui.setupUi(self)
-
-        graph_phasing = self.ui.PhasingGraph
-        graph_phasing.getAxis('bottom').setLabel("Frequency, MHz")
-        graph_phasing.getAxis('left').setLabel("Amplitude, a.u.")
-        graph_phasing.setTitle("Phasing")
-        self.ui.PhasingGraph.addLegend()
-
-        self.ui.pushButton_2.clicked.connect(self.zero)
-        self.ui.verticalSlider_a.valueChanged.connect(self.value_changed)
-        self.ui.verticalSlider_b.valueChanged.connect(self.value_changed)
-        self.ui.verticalSlider_c.valueChanged.connect(self.value_changed)
-        self.ui.verticalSlider_d.valueChanged.connect(self.value_changed)
-        self.ui.dial.valueChanged.connect(self.smoothing_changed)
-        self.ui.pushButton_3.clicked.connect(self.manual_read)
-        self.ui.pushButton.clicked.connect(self.save_data)
-
-        self.Real_freq_phased = None
-
-    def read_data(self):
-        global Frequency, Re_spectra, Im_spectra
-
-        self.zero()
-
-    def save_data(self):
-        global Re_spectra
-
-        if self.Real_freq_phased is not None:
-            Re_spectra = self.Real_freq_phased
-        else:
-            Re_spectra = Re_spectra
-
-        self.close()
-
-    def closeEvent(self, event):
-        self.closed.emit()
-        super().closeEvent(event)
-
-    def zero(self):
-        self.a = 0
-        self.b = 0
-        self.c = 0
-        self.d = 0
-        self.Smooth = 0
-
-        self.set_zero(self.ui.verticalSlider_a, self.ui.Box_a)
-        self.set_zero(self.ui.verticalSlider_b, self.ui.Box_b)
-        self.set_zero(self.ui.verticalSlider_c, self.ui.Box_c)
-        self.set_zero(self.ui.verticalSlider_d, self.ui.Box_d)
-
-        self.ui.dial.setValue(0)
-        self.ui.Box_smooth.setValue(0)
-
-        if self.Real_freq_phased is not None:
-            self.process_data()
-
-    def set_zero(self,slider,box):
-        slider.setValue(0)
-        box.setValue(0)
-
-    def process_data(self):
-        self.Real_freq_phased = self.calculate_phase()
-        self.update_plot()
-        self.update_text()
-
-    def calculate_phase(self):
-        global Frequency, Re_spectra, Im_spectra
-        phi = self.a + self.b * Frequency + self.c * Frequency ** 2 + self.d * Frequency ** 3
-        self.Real_freq_phased = Re_spectra * np.cos(np.deg2rad(phi)) - Im_spectra * np.sin(np.deg2rad(phi))
-
-        if self.Smooth > 1:
-            self.Smooth = int(self.Smooth)
-            self.Real_freq_phased = self.smooth(self.Real_freq_phased)
-
-        return self.Real_freq_phased
-
-    def smooth(self,y,):
-        new_array = savgol_filter(y, window_length=self.Smooth, polyorder=1)
-        # print(self.Smooth)
-
-        return new_array
-
-    def update_plot(self):
-        global Frequency, Re_spectra, Im_spectra
-
-        self.ui.PhasingGraph.clear()
-        self.ui.PhasingGraph.plot(Frequency, Re_spectra, pen='r', name = 'Original') #Real origin
-        self.ui.PhasingGraph.plot(Frequency, self.Real_freq_phased, pen='b', name = 'Phased') #Real phased
-
-    def update_text(self):
-        global Frequency
-        Integral = np.trapz(self.Real_freq_phased)
-        left_mean = np.mean(self.Real_freq_phased[:100])
-        right_mean = np.mean(self.Real_freq_phased[-100:])
-        delta = left_mean - right_mean
-
-        self.ui.Integral.setText(f"Integral: {round(Integral,3)}")
-        self.ui.Delta.setText(f"Delta: {round(delta,7)}")
-
-        Real_apod   = Cal._calculate_apodization(self.Real_freq_phased, Frequency)
-        M2, T2 = Cal._calculate_M2(Real_apod, Frequency)
-
-        self.ui.M2.setText(f"M₂: {round(M2,5)}")
-        self.ui.T2.setText(f"T₂*: {round(T2,3)}")
-
-    def value_changed(self):
-        global Frequency, Re_spectra, Im_spectra
-
-        self.a = self.ui.verticalSlider_a.value()
-        self.b = self.ui.verticalSlider_b.value()
-        self.c = self.ui.verticalSlider_c.value()
-        self.d = self.ui.verticalSlider_d.value()
-
-        self.ui.Box_a.setValue(self.a)
-        self.ui.Box_b.setValue(self.b)
-        self.ui.Box_c.setValue(self.c)
-        self.ui.Box_d.setValue(self.d)
-
-        self.check_borders(self.ui.verticalSlider_a)
-        self.check_borders(self.ui.verticalSlider_b)
-        self.check_borders(self.ui.verticalSlider_c)
-        self.check_borders(self.ui.verticalSlider_d)
-
-
-        if Frequency is not None and Re_spectra is not None and Im_spectra is not None:
-            self.process_data()
-        else:
-            return
-
-    def check_borders(self, slider):
-        max = slider.maximum()
-        min = slider.minimum()
-
-        val = slider.value()
-
-        if val + 10 > max:
-            slider.setMaximum(val+10)
-        elif val - 10 < min:
-            slider.setMinimum(val-10)
-
-    def smoothing_changed(self):
-        self.Smooth = self.ui.dial.value()
-        self.ui.Box_smooth.setValue(self.Smooth)
-        self.process_data()
-
-    def manual_read(self):
-        self.a = self.ui.Box_a.value()
-        self.b = self.ui.Box_b.value()
-        self.c = self.ui.Box_c.value()
-        self.d = self.ui.Box_d.value()
-        self.Smooth = self.ui.Box_smooth.value()
-
-        self.ui.verticalSlider_a.valueChanged.disconnect(self.value_changed)
-        self.ui.verticalSlider_b.valueChanged.disconnect(self.value_changed)
-        self.ui.verticalSlider_c.valueChanged.disconnect(self.value_changed)
-        self.ui.verticalSlider_d.valueChanged.disconnect(self.value_changed)
-        self.ui.dial.valueChanged.disconnect(self.smoothing_changed)
-
-        self.ui.dial.setValue(self.Smooth)
-        self.ui.verticalSlider_a.setValue(self.a)
-        self.ui.verticalSlider_b.setValue(self.b)
-        self.ui.verticalSlider_c.setValue(self.c)
-        self.ui.verticalSlider_d.setValue(self.d)
-
-        self.process_data()
-
-        self.ui.verticalSlider_a.valueChanged.connect(self.value_changed)
-        self.ui.verticalSlider_b.valueChanged.connect(self.value_changed)
-        self.ui.verticalSlider_c.valueChanged.connect(self.value_changed)
-        self.ui.verticalSlider_d.valueChanged.connect(self.value_changed)
-        self.ui.dial.valueChanged.connect(self.smoothing_changed)
